requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1

# Testing dependencies
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0